[Bronze V] Title: 이상한 기호, Time: 204 ms, Memory: 17752 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -4,8 +4,8 @@ import java.util.*;
 public class Main {
 	public static void main(String[] args) throws IOException {
 		Scanner sc = new Scanner(System.in);
-		int A = sc.nextInt();
-		int B = sc.nextInt();
+		long A = sc.nextLong();
+		long B = sc.nextLong();
 		sc.close();
 		System.out.println((A + B) * (A - B));
 	}
